Remove CStore
