--- src/js/Content/Features/Store/Common/CStore.js
@@ -1,14 +1,0 @@
-import {CStoreBase} from "./CStoreBase";
-
-export class CStore extends CStoreBase {
-
-    getAllSubids() {
-        const result = [];
-        for (const node of document.querySelectorAll("input[name=subid]")) {
-            if (node.value) {
-                result.push(node.value);
-            }
-        }
-        return result;
-    }
-}
