+++ src/js/Content/Features/Store/Sub/CSub.js
@@ -1,6 +1,6 @@
 import ContextType from "../../../Modules/Context/ContextType";
 import {GameId} from "../../../../Core/GameId";
-import {CStore} from "../Common/CStore";
+import {CStoreBase} from "../Common/CStoreBase";
 import FExtraLinks from "../Common/FExtraLinks";
 import FDRMWarnings from "../Common/FDRMWarnings";
 import FITADPrices from "../Common/FITADPrices";
@@ -8,7 +8,7 @@ import FRegionalPricing from "../Common/FRegionalPricing";
 import FAddToCartNoRedirect from "../Common/FAddToCartNoRedirect";
 import FSavingsCheck from "./FSavingsCheck";
 
-export class CSub extends CStore {
+export class CSub extends CStoreBase {
 
     constructor() {
